Checkpath os-specific logic so godel can build on windows.
