+++ cmd/checkpath/checkpath.go
@@ -23,7 +23,6 @@ import (
 	"regexp"
 	"runtime"
 	"strings"
-	"syscall"
 
 	"github.com/pkg/errors"
 	"github.com/termie/go-shutil"
@@ -167,31 +166,6 @@ func checkForCfgFile(name string) string {
 		return cfgFile
 	}
 	return ""
-}
-
-func pathsOnSameDevice(p1, p2 string) bool {
-	id1, ok := getDeviceID(p1)
-	if !ok {
-		return false
-	}
-	id2, ok := getDeviceID(p2)
-	if !ok {
-		return false
-	}
-	return id1 == id2
-}
-
-func getDeviceID(p string) (interface{}, bool) {
-	fi, err := os.Stat(p)
-	if err != nil {
-		return 0, false
-	}
-	s := fi.Sys()
-	switch s := s.(type) {
-	case *syscall.Stat_t:
-		return s.Dev, true
-	}
-	return 0, false
 }
 
 func gitRepoRootPath(dir string) (string, error) {
